pyproject.toml
[tool.poetry]
name = "headless-cms"
version = "0.1.0"
description = ""
authors = ["Mohammed <mohammed.mm404@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-httpauth = "^4.8.0"
jwt = "^1.3.1"


[tool.poetry.group.dev.dependencies]
httpie = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
